requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
bcrypt==4.1.2
python-dotenv==1.0.0
google-generativeai==0.6.0
reportlab==4.2.5
psycopg2-binary==2.9.11
Flask-Migrate==4.0.7
opencv-python==4.8.1.78
numpy>=1.24.0
sentence-transformers==2.2.2
language-tool-python==2.7.1
nltk==3.8.1
deepface==0.0.79
pandas>=2.0.0,<3.0.0
tensorflow>=2.13.0,<3.0.0
setuptools>=65.0.0
spacy==3.7.2

